refactor: keep audit-touched modules under their tier caps
Extract the label-fold vocabularies, the Postgres brain marshalling adapters, and the docs-engine record mappers into cohesive sibling modules so prometheus_labels, project_brain_repo, and docs_engine/service stay under the module-size budget, and drop historical framing from the currency-resolver docstring.
